requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.1
